Redistribute file system
We have thought that separating the core of the app framework will
help to maintain a version control that is separate and cleaner compared
to the previous versions. This will allow developers to have two version
controls.

One with the latest updates from the Cavesman and another from our project,
without affecting Cavesman's own nucleus